move queue test t its own file
